Performance optimizations and consolidation of all models into a single model type
